requirements.txt
# PDF parsing
llama-parse

# Core RAG
langchain
langchain-community
langchain-openai
langgraph

# Vector + retrieval
pinecone-client
rank-bm25

# Reranking
cohere

# Embeddings
tiktoken

# Evaluation
ragas
datasets

# API + UI
fastapi
uvicorn
pydantic
streamlit
requests

# Database
psycopg2-binary

# Bot
python-telegram-bot

# Monitoring
langfuse

# Utils
python-dotenv
openai
pypdf
pandas
numpy

# MCP
mcp
fastmcp
